pyproject.toml
[tool.poetry]
name = "aws-3tier-platform"
version = "0.1.0"
description = "Production-style 3-tier AWS platform"
authors = ["simoda"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
aws-cdk-lib = "^2.170.0"
constructs = "^10.0.0"
boto3 = "^1.34.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
